fix(linter): react/jsx-fragments rule should take string argument.
The original rule has a config option in the format of `["error", "element"]` or `["error", "syntax"]`. We required a config object, which is technically incorrect.

This fixes that in the docs, but we still need to fix the code to accept a string argument instead of an object. I guess we have to make this backward-compatible and accept either?